move constants to constants folder
